+++ src/legacy/ui/public/visualize/loader/pipeline_helpers/build_pipeline.ts
@@ -23,7 +23,7 @@ import { setBounds } from 'ui/agg_types/buckets/date_histogram';
 import { SearchSource } from 'ui/courier';
 import { AggConfig, Vis, VisParams, VisState } from 'ui/vis';
 import moment from 'moment';
-import { SerializedFieldFormat } from 'src/plugins/data/common';
+import { SerializedFieldFormat } from 'src/plugins/data/common/expressions/types/common';
 import { createFormat } from './utilities';
 
 interface SchemaConfigParams {
